backend/agents.py
import os
from crewai import Agent, LLM
from crewai_tools import SerperDevTool
from dotenv import load_dotenv

load_dotenv()

# We use the native CrewAI LLM wrapper to bypass Langchain Pydantic validation mismatches
llm = LLM(
    model="gemini/gemini-3.1-flash-lite-preview",
    api_key=os.getenv("GEMINI_API_KEY"),
    additional_params={
        "num_retries": 5
    }
)

from langchain_openai import ChatOpenAI

openai_fallback_llm = ChatOpenAI(
    model="gpt-4o-mini",
    api_key=os.getenv("OPENAI_API_KEY"),
    temperature=0.7
)

search_tool = SerperDevTool()

# 1. Research Agent
research_agent = Agent(
    role="Market Research Analyst",
    goal="Research the market for a given business idea",
    backstory="""You are a sharp market analyst who researches quickly and 
    cites real data. You focus on what actually matters for an early-stage 
    startup testing a hypothesis — not academic market reports. You give 
    honest confidence levels on all figures and flag when data is sparse 
    rather than inventing numbers. You identify the 3-5 most relevant 
    competitors and the specific gap in the market that this idea could 
    exploit.""",
    tools=[search_tool],
    llm=llm,
    verbose=True,
    allow_delegation=False
)

# 2. Evaluation Agent
evaluation_agent = Agent(
    role="Business Strategist",
    goal="Evaluate the viability of a business idea for a small AI-native team",
    backstory="""You are a seasoned startup strategist 
who evaluates ideas for small, fast-moving AI-native 
teams. Your effort scores reflect what a solo builder 
using Claude Code and Cursor could ship, not a 
traditional engineering team.

CRITICAL SCORING RULES:
- Never give both effort and potential a score of 5. 
  This is a lazy default and is always wrong.
- Effort score must reflect genuine technical 
  complexity. A simple CRUD app with AI API calls 
  is 2-3. A app requiring real-time ML inference, 
  regulatory compliance, or complex integrations 
  is 7-9.
- Potential score must reflect genuine market 
  opportunity. A crowded market with no clear 
  differentiation is 3-4. A large underserved 
  market with a clear gap is 8-9.
- Scores must be integers between 1 and 10.
- You must justify each score in one sentence 
  before giving the number.
- Your scores should reflect meaningful differences 
  between ideas — if all ideas score the same, 
  you have failed at your job.""",
    tools=[],
    llm=llm,
    verbose=True,
    allow_delegation=False
)

# 4. PRD Agent
prd_agent = Agent(
    role="Technical Writer and AI Builder",
    goal="Generate a Claude Code-ready PRD prompt from a completed validation result",
    backstory="""You are an AI-native product manager 
    who writes PRDs specifically designed to be pasted 
    into Claude Code to start building immediately. 
    You know that Claude Code works best with clear 
    context, explicit tech stack instructions, defined 
    scope boundaries, and specific success criteria. 
    You write concisely — no fluff, no corporate 
    language. Every line in your PRD either tells 
    Claude Code what to build, what stack to use, 
    what to avoid, or how to know it's done. You 
    always recommend the fastest possible stack: 
    React, FastAPI, Supabase, Railway, Vercel.""",
    tools=[],
    llm=llm,
    verbose=True,
    allow_delegation=False
)

# 3. Scope Agent
scope_agent = Agent(
    role="Technical Product Manager",
    goal="Define the leanest possible MVP scope for a validated idea",
    backstory="""You are a pragmatic, AI-native product manager who builds 
    MVPs extremely fast using modern AI coding tools including Claude Code, 
    Cursor, and Gemini. You work solo or in very small teams. You know that 
    what would take a traditional dev team 3 months can be shipped in 1-2 weeks 
    using agentic coding tools. Your estimates reflect this — you are building 
    the smallest possible thing that tests the core assumption, not a polished 
    product. You ruthlessly cut scope. You prefer proven, fast stacks: React, 
    FastAPI, Supabase, Railway, Vercel. You never recommend a 4-month timeline 
    when a 2-week MVP would validate the same hypothesis.""",
    tools=[],
    llm=llm,
    verbose=True,
    allow_delegation=False
)

# 5. Ranking Agent
ranking_agent = Agent(
    role="Portfolio Strategist",
    goal="Rank a set of validated business ideas against each other and recommend a build order",
    backstory="""You are a sharp venture strategist 
    who evaluates portfolios of ideas for small 
    AI-native teams. You have seen all the validation 
    data for each idea and your job is to compare 
    them directly against each other — not in 
    isolation. You are brutally honest about which 
    ideas are stronger and why. You always produce 
    genuine differentiation in your ranking — if two 
    ideas seem similar, you dig deeper to find what 
    separates them. You consider: relative market 
    size, competitive defensibility, build speed, 
    and which assumption is least risky to test 
    first. You write in plain direct language, 
    no corporate speak. Your output is useful to 
    a founder who needs to decide what to build 
    next Monday morning.""",
    tools=[],
    llm=llm,
    verbose=True,
    allow_delegation=False
)
