backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
crewai
crewai-tools
google-generativeai
langchain-google-genai
langchain-openai
